+++ e2e/react-router/basic-esbuild-file-based/playwright.config.ts
@@ -1,12 +1,8 @@
 import { defineConfig, devices } from '@playwright/test'
-import {
-  getDummyServerPort,
-  getTestServerPort,
-} from '@tanstack/router-e2e-utils'
-import packageJson from './package.json' with { type: 'json' }
+import { appServerReady } from '@tanstack/router-e2e-utils'
 
-const PORT = await getTestServerPort(packageJson.name)
-const EXTERNAL_PORT = await getDummyServerPort(packageJson.name)
+const PORT = Number(process.env.E2E_APP_PORT ?? 0)
+
 const baseURL = `http://localhost:${PORT}`
 /**
  * See https://playwright.dev/docs/test-configuration.
@@ -17,18 +13,15 @@ export default defineConfig({
 
   reporter: [['line']],
 
-  globalSetup: './tests/setup/global.setup.ts',
-  globalTeardown: './tests/setup/global.teardown.ts',
-
   use: {
     /* Base URL to use in actions like `await page.goto('/')`. */
     baseURL,
   },
 
   webServer: {
-    command: `pnpm run build && pnpm run serve --serve=${PORT} --define:process.env.NODE_ENV=\\"test\\" --define:process.env.EXTERNAL_PORT="${EXTERNAL_PORT}"`,
-    url: baseURL,
-    reuseExistingServer: !process.env.CI,
+    command: `pnpm run build && pnpm run serve --serve=${PORT} --define:process.env.NODE_ENV=\\"test\\"`,
+    wait: appServerReady,
+    reuseExistingServer: false,
     stdout: 'pipe',
   },