+++ e2e/vue-router/basic-vue-query-file-based/src/posts.tsx
@@ -8,11 +8,7 @@ export type PostType = {
 
 export class PostNotFoundError extends Error {}
 
-let queryURL = 'https://jsonplaceholder.typicode.com'
-
-if (import.meta.env.VITE_NODE_ENV === 'test') {
-  queryURL = `http://localhost:${import.meta.env.VITE_EXTERNAL_PORT}`
-}
+const queryURL = 'https://jsonplaceholder.typicode.com'
 
 export const fetchPost = async (postId: string) => {
   console.info(`Fetching post with id ${postId}...`)