+++ e2e/vue-start/basic/src/routes/api/users.$userId.ts
@@ -1,19 +1,15 @@
 import { createFileRoute } from '@tanstack/vue-router'
 import type { User } from '~/utils/users'
 
-let queryURL = 'https://jsonplaceholder.typicode.com'
-
-if (import.meta.env.VITE_NODE_ENV === 'test') {
-  queryURL = `http://localhost:${import.meta.env.VITE_EXTERNAL_PORT}`
-}
-
 export const Route = createFileRoute('/api/users/$userId')({
   server: {
     handlers: {
       GET: async ({ params, request }) => {
         console.info(`Fetching users by id=${params.userId}... @`, request.url)
         try {
-          const res = await fetch(`${queryURL}/users/${params.userId}`)
+          const res = await fetch(
+            `https://jsonplaceholder.typicode.com/users/${params.userId}`,
+          )
           if (!res.ok) {
             throw new Error('Failed to fetch user')
           }