+++ e2e/vue-start/basic/src/routes/redirect/$target/index.tsx
@@ -1,9 +1,10 @@
+import { defineComponent } from 'vue'
 import { Link, createFileRoute } from '@tanstack/vue-router'
 
-export const Route = createFileRoute('/redirect/$target/')({
-  component: () => {
+const RouteComponent = defineComponent({
+  setup() {
     const preload = Route.useSearch({ select: (s) => s.preload })
-    return (
+    return () => (
       <div>
         <div class="mb-2">
           <Link
@@ -73,4 +74,8 @@ export const Route = createFileRoute('/redirect/$target/')({
       </div>
     )
   },
+})
+
+export const Route = createFileRoute('/redirect/$target/')({
+  component: RouteComponent,
 })